move condition type to conditional.js
