+++ src/utils.js
@@ -1,4 +1,7 @@
-module.exports= {getField:getField};
+module.exports={
+    getField:getField,
+    evalCondition:evalCondition
+};
 
 function getField(countField, rootNode) {
     var countFieldArr = countField.split(".");
@@ -7,4 +10,15 @@ function getField(countField, rootNode) {
         count = count[countFieldArr[index]];
     }
     return count;
+}
+
+
+function evalCondition(condition,field_values)
+{
+    var field_value_to_test="this" in condition && condition["this"] ? field_values["this"][condition.field] : field_values[condition.field];
+    var b=condition.values.some(function(value) {return field_value_to_test===value;});
+    if("different" in condition && condition["different"])
+        return !b;
+    else
+        return b;
 }
